+++ JMS-Test/src/main/java/com/ncr/ATMMonitoring/serverchain/message/specific/incoming/UpdateDataResponse.java
@@ -45,7 +45,7 @@ public class UpdateDataResponse implements SpecificMessage{
         return originalRequest.getAtmIp();
     }
 
-    public int getMatricula() {
+    public Long getMatricula() {
         return originalRequest.getMatricula();
     }
